Fixed unit tests (Travis may still complain) and many bugs causing unit test failures. Implemented JSON migration to sqlite3

+++ continuousprint/storage/queries.py
@@ -8,12 +8,19 @@ from .database import Queue, Job, Set, Run, DB, DEFAULT_QUEUE, ARCHIVE_QUEUE
 
 
 def getQueues():
-    return Queue.select().order_by(Queue.lexRank.asc())
+    return Queue.select().order_by(Queue.rank.asc())
 
 
 def getAcquired():
-    # TODO fetch job, set, and run
+    r = Run.select().where(Run.end.is_null()).limit(1).execute()
+    if len(r) > 0:
+        r = r[0]
+        j = r.job
+        for s in j.sets:
+            if s.remaining > 0 and s.path == r.path:
+                return (j, s, r)
     return (None, None, None)
+
 
 def assignQueues(queues):
     # Default/archive queues should never be removed
@@ -30,74 +37,68 @@ def assignQueues(queues):
             absent_names = []
         added = [q for q in queues if q["name"] not in qq_names]
         added_names = set([q["name"] for q in added])
-        for lex, qdata in enumerate(queues):
+        for rank, qdata in enumerate(queues):
             if qdata["name"] in added_names:
                 print("Create", qdata)
                 Queue.create(
                     name=qdata["name"],
                     strategy=qdata["strategy"],
                     addr=qdata["addr"],
-                    lexRank=lex,
+                    rank=rank,
                 )
             else:
                 print("Update", qdata)
                 Queue.update(
-                    strategy=qdata["strategy"], addr=qdata["addr"], lexRank=lex
+                    strategy=qdata["strategy"], addr=qdata["addr"], rank=rank
                 ).where(Queue.name == qdata["name"]).execute()
     return (absent_names, added)
 
 
 def getJobsAndSets(queue):
     if type(queue) == str:
-        queue = Queue.get(id=queue)
-    cursor = (
+        queue = Queue.get(name=queue)
+    return (
         Job.select()
         .join(Set, JOIN.LEFT_OUTER)
-        .where(Job.queue == q)
+        .where(Job.queue == queue)
         .group_by(Job.id)
-        .order_by(Job.lexRank.asc())
-    )
-    return cursor.execute()
+        .order_by(Job.rank.asc())
+    ).execute()
 
 
 def getJob(jid):
     return Job.get(id=jid)
 
 
-def getJobsAndSets(q=None, lexOrder=False):
-    if type(q) == str:
-        q = Queue.get(name=q)
-    if q is None:
-        return []
-    cursor = Job.select().join(Set, JOIN.LEFT_OUTER).where(Job.queue == q)
-    cursor = cursor.group_by(Job.id)
-    if lexOrder:
-        cursor = cursor.order_by(Job.lexRank.asc())
-    return cursor.execute()
-
-
-def decrementJobRemaining(j):
-    j.remaining = max(j.remaining - 1, 0)
-    j.save()
-    if j.remaining > 0:
-        # Refresh sets within the job if this isn't the last go-around
-        Set.update(remaining=Set.count).where(Set.job == j).execute()
-        return True
-    else:
-        return False
-
-
 def getNextJobInQueue(q=None):
     # Need to loop over jobs first to maintain job order
-    for job in getJobsAndSets(q=q, lexOrder=True):
-        for set_ in job.sets:
-            if set_.remaining > 0:
-                return job
-
-        # If we've looped through all sets in the job and all are complete,
-        # decrement the job. Pick its first set if there's still more work to be done.
-        if job.remaining > 0 and decrementJobRemaining(job):
+    for job in getJobsAndSets(q):
+        has_work = job.normalize()
+        if has_work:
             return job
+
+
+def _updateSet(set_id, data, job):
+    # Called internally from updateJob
+    s = Set.get(id=set_id)
+    for k, v in data.items():
+        if k in ("id", "count", "remaining"):  # ignored or handled below
+            continue
+        setattr(s, k, v)
+    s.job = job
+
+    if data.get("count") is not None:
+        newCount = int(data["count"])
+        inc = newCount - s.count
+        s.count = newCount
+        s.remaining = min(newCount, s.remaining + max(inc, 0))
+        # Boost job remaining if we would cause it to be incomplete
+        job_remaining = s.job.remaining
+        if inc > 0 and job_remaining == 0:
+            job_remaining = 1
+            s.job.remaining = 1
+            s.job.save()
+    s.save()
 
 
 def updateJob(job_id, data, queue=DEFAULT_QUEUE):
@@ -131,113 +132,107 @@ def updateJob(job_id, data, queue=DEFAULT_QUEUE):
         return Job.get(id=job_id).as_dict()
 
 
-MAX_LEX = 1000000.0  # Arbitrary
+MAX_RANK = 1000000.0  # Arbitrary
 
 
-def genLex(n):
-    maxval = MAX_LEX
+def _genRank(n):
+    maxval = MAX_RANK
     stride = int(maxval / (n + 1))  # n+1 to allow for space at beginning
     for i in range(stride, int(maxval + 1), stride):
         yield i
 
 
-def lexBalance():
+def _rankBalance(cls):
     with DB.queues.atomic():
         # TODO discriminate by queue - may case weirdness with archived jobs
-        lexer = genLex(Job.select().count())
-        for (l, c) in zip(lexer, Job.select().order_by(Job.lexRank)):
-            c.lexRank = l
+        ranker = _genRank(cls.select().count())
+        for (l, c) in zip(ranker, cls.select().order_by(cls.rank)):
+            c.rank = l
             c.save()
 
 
-def lexEnd():
+def _rankEnd():
     return time.time()
 
 
-def moveJob(src_id: int, dest_id: int, retried=False):
+def _moveImpl(cls, src, dest_id: int, retried=False):
     if dest_id == -1:
         destRank = 0
     else:
-        destRank = Job.get(id=dest_id).lexRank
-    # Get the next object/set having a lexRank beyond the destination rank,
+        destRank = cls.get(id=dest_id).rank
+
+    # Get the next object having a rank beyond the destination rank,
     # so we can then split the difference
     # Note the unary '&' operator and the expressions wrapped in parens (a limitation of peewee)
     postRank = (
-        Job.select(Job.lexRank)
-        .where((Job.lexRank > destRank) & (Job.id != src_id))
+        cls.select(cls.rank)
+        .where((cls.rank > destRank) & (cls.id != src.id))
         .limit(1)
         .execute()
     )
     if len(postRank) > 0:
-        postRank = postRank[0].lexRank
+        postRank = postRank[0].rank
     else:
-        postRank = MAX_LEX
-    # Pick the target value as the midpoint between the two lexRanks
+        postRank = MAX_RANK
+    # Pick the target value as the midpoint between the two ranks
     candidate = abs(postRank - destRank) / 2 + min(postRank, destRank)
 
     # We may end up with an invalid candidate if we hit a singularity - in this case, rebalance all the
     # rows and try again
     if candidate <= destRank or candidate >= postRank:
         if not retried:
-            lexBalance()
-            moveJob(src_id, dest_id, retried=True)
+            _rankBalance(cls)
+            _moveImpl(cls, src, dest_id, retried=True)
         else:
-            raise Exception("Could not rebalance job lexRank to move job")
+            raise Exception("Could not rebalance job rank to move job")
     else:
-        c = Job.get(id=src_id)
-        c.lexRank = candidate
-        c.save()
+        src.rank = candidate
+        src.save()
 
 
-def newEmptyJob(q, lex=lexEnd):
+def moveJob(src_id: int, dest_id: int):
+    j = Job.get(id=src_id)
+    return _moveImpl(Job, j, dest_id)
+
+
+def moveSet(src_id: int, dest_id: int, dest_job: int):
+    s = Set.get(id=src_id)
+    if dest_job == -1:
+        j = newEmptyJob(s.job.queue)
+    else:
+        j = Job.get(id=dest_job)
+    s.job = j
+    s.save()
+    _moveImpl(Set, s, dest_id)
+
+
+def newEmptyJob(q, name="", rank=_rankEnd):
     if type(q) == str:
         q = Queue.get(name=q)
     j = Job.create(
         queue=q,
-        lexRank=lex(),
-        name="",
+        rank=rank(),
+        name=name,
         count=1,
     )
     return j
 
 
-def importJob(q, data, lex=lexEnd):
-    if type(q) == str:
-        q = Queue.get(name=q)
-    with DB.queues.atomic():
-        j = Job.create(
-            queue=q,
-            draft=False,
-            lexRank=lex(),
-            name=data["name"],
-            count=data["count"],
-            remaining=data["count"],
-        )
-        for s in data["sets"]:
-            Set.create(
-                path=s["path"],
-                sd=False,
-                job=j,
-                lexRank=lex(),
-                count=s["count"],
-                remaining=s["count"],
-                material_keys=(",".join(s["materials"])),
-            )
-    return j
-
-
-def appendSet(queue: str, job, data: dict, lex=lexEnd):
+def appendSet(queue: str, job, data: dict, rank=_rankEnd):
     q = Queue.get(name=queue)
     if job == "":
         j = newEmptyJob(q)
     else:
-        j = Job.get(queue=q, name=job)
+        try:
+            j = Job.get(queue=q, name=job)
+        except Job.DoesNotExist:
+            j = newEmptyJob(q, name=job)
 
     count = int(data["count"])
     s = Set.create(
         path=data["path"],
         sd=data["sd"] == "true",
-        lexRank=lex(),
+        rank=rank(),
         material_keys=",".join(data["material"]),
         count=count,
         remaining=count,
@@ -247,43 +242,28 @@ def appendSet(queue: str, job, data: dict, lex=lexEnd):
     return dict(job_id=j.id, set_=s.as_dict())
 
 
-def _updateSet(set_id, data, job):
-    s = Set.get(id=set_id)
-    for k, v in data.items():
-        if k in ("id", "count", "remaining"):  # ignored or handled below
-            continue
-        setattr(s, k, v)
-    s.job = job
-
-    if data.get("count") is not None:
-        newCount = int(data["count"])
-        inc = newCount - s.count
-        s.count = newCount
-        s.remaining = min(newCount, s.remaining + max(inc, 0))
-        # Boost job remaining if we would cause it to be incomplete
-        job_remaining = s.job.remaining
-        if inc > 0 and job_remaining == 0:
-            job_remaining = 1
-            s.job.remaining = 1
-            s.job.save()
-    s.save()
-
-
-def removeQueues(queue_ids: list):
-    return {"queues_deleted": Queue.delete().where(Queue.id.in_(queue_ids)).execute()}
-
-
-def removeJobsAndSets(job_ids: list, set_ids: list):
-    q = Queue.get(name="archive")
+def remove(queue_ids: list[int] = [], job_ids: list[int] = [], set_ids: list[int] = []):
+    result = {}
     with DB.queues.atomic():
-        j = Job.update(queue=q).where(Job.id.in_(job_ids)).execute()
+        if len(queue_ids) > 0:
+            result["queues_deleted"] = (
+                Queue.delete().where(Queue.id.in_(queue_ids)).execute()
+            )
+
+        q = Queue.get(name="archive")
+
+        if len(job_ids) > 0:
+            result["jobs_deleted"] = (
+                Job.update(queue=q).where(Job.id.in_(job_ids)).execute()
+            )
         # Only delete sets if we haven't already archived their job
-        s = (
-            Set.delete()
-            .where((Set.id.in_(set_ids)) & (Set.job.not_in(job_ids)))
-            .execute()
-        )
-    return {"jobs_deleted": j, "sets_deleted": s}
+        if len(set_ids) > 0:
+            result["sets_deleted"] = (
+                Set.delete()
+                .where((Set.id.in_(set_ids)) & (Set.job.not_in(job_ids)))
+                .execute()
+            )
+    return result
 
 
 def replenish(job_ids: list, set_ids: list):
@@ -306,7 +286,10 @@ def beginRun(s):
     Run.update({Run.end: datetime.now(), Run.result: "aborted"}).where(
         (Run.job == s.job) & (Run.end.is_null())
     ).execute()
-    return Run.create(job=s.job, path=s.path)  # start defaults to now()
+    j = s.job
+    return Run.create(
+        queueName=j.queue.name, jobName=j.name, job=s.job, path=s.path
+    )  # start defaults to now()
 
 
 def endRun(s, r, result: str, txn=None):
@@ -330,7 +313,7 @@ def endRun(s, r, result: str, txn=None):
                     .count
                 )
                 if totalRemaining == 0:
-                    decrementJobRemaining(j)
+                    j.decrement(save=True)
 
 
 def getHistory():
